central healthcheck, kafka localhost listener
Added healthcheck to central so that drivers and monitors know when they can start trying to connect to central. It required "curl" to ping the localhost:8000/health so i added it to dockerfile for the central.

Also added new listening port for kafka to be able to look into content of kafka from cli