creating a ticket in the engine main
The code creates a ticket in stop_suplly(), stores it locally and tries to send it to the Central with Kafka.
But be aware there is a problem in this code.
It uses new kafka topic which was never created - and I am not sure if a new kafka topic is actually necessary.
I need to look into it.